zmq_default/requirements.txt
pyzmq
python-dotenv
Flask
requests
